User Crude completed

+++ DataAccessLayer/User.cs
@@ -34,7 +34,7 @@ namespace DataAccessLayer
         public string Token { get; set; }
         public long MobileNo { get; set; }
         public string ImageString { get; set; }
-        public Nullable<byte> UsersStatus { get; set; }
+        public Nullable<int> UsersStatus { get; set; }
         public Nullable<System.DateTime> DateCreated { get; set; }
         public Nullable<System.DateTime> DateUpdated { get; set; }
